move the avro schemas off the flink namespace
