+++ examples/taxi-ride/datamodel/src/main/avro/taxiride.avsc
@@ -1,5 +1,5 @@
 {
-    "namespace": "taxiride.flink.avro",
+    "namespace": "taxiride.datamodel",
     "type": "record",
     "name": "TaxiRide",
     "fields":[
