+++ examples/taxi-ride/ingestor/src/main/scala/taxiride/ingestor/JsonFormats.scala
@@ -17,7 +17,7 @@
 package taxiride.ingestor
 
 import spray.json._
-import taxiride.flink.avro._
+import taxiride.datamodel._
 
 object TaxiRideJsonProtocol extends DefaultJsonProtocol {
   implicit object TaxiRideJsonFormat extends RootJsonFormat[TaxiRide] {
