+++ examples/taxi-ride/logger/src/main/scala/taxiride/logger/FarePerRideLogger.scala
@@ -20,7 +20,7 @@ import cloudflow.akkastream._
 import cloudflow.akkastream.scaladsl._
 import cloudflow.streamlets._
 import cloudflow.streamlets.avro._
-import taxiride.flink.avro._
+import taxiride.datamodel._
 
 class FarePerRideLogger extends AkkaStreamlet {
   val inlet = AvroInlet[TaxiRideFare]("in")
